+++ R/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(statcanR)
+
+test_check("statcanR")